+++ wheelly/renders.py
@@ -1,11 +1,10 @@
 import math
-from typing import Iterable
+from typing import Iterable, Tuple
 import pygame
 import numpy as np
 from numpy import ndarray
-from pygame import Surface
 
-from wheelly.robot import RobotAPI
+from wheelly.robots import RobotAPI
 
 WINDOW_SIZE = 800
 GRID_SIZE = 0.2
@@ -41,6 +40,9 @@ SENSOR_SHAPE = np.array([
     (SENSOR_LENGTH, 0.0)
 ])
 
+_FONT_NAME = 'freesans'
+_FONT_SIZE = 16
+
 class RobotWindow:
 
     def __init__(self):
@@ -50,7 +52,10 @@ class RobotWindow:
         self._robor_pos= np.zeros(2)
         self._robor_dir= 0
         self._sensor_dir = 0
-    
+        self._font = pygame.font.SysFont(_FONT_NAME, _FONT_SIZE) 
+        self._canvas = pygame.Surface((WINDOW_SIZE, WINDOW_SIZE))
+        self._trans = _transMatrix()
+
     def robot_pos(self, robot_location:ndarray):
         self._robot_pos = robot_location
         return self
@@ -67,58 +72,94 @@ class RobotWindow:
         self._robot_pos = robot.robot_pos()
         self._robot_dir = robot.robot_dir()
         self._sensor_dir = robot.sensor_dir()
+        self._distance = robot.sensor_distance()
+        self._sensor_obstacle = robot.sensor_obstacle()
+        self._obstaclesMap = robot.obstaclesMap()
         return self
 
     def render(self):
-        canvas = pygame.Surface((WINDOW_SIZE, WINDOW_SIZE))
-        canvas.fill(BACKGROUND_COLOR)
-        trans = transMatrix()
-        drawRobot(canvas, trans, self._robot_pos, self._robot_dir)
-        drawSensor(canvas, trans, self._robot_pos, self._robot_dir, self._sensor_dir)
+        self._canvas.fill(BACKGROUND_COLOR)
+        self._drawRobot() \
+            ._drawSensor() \
+            ._drawObstacles() \
+            ._drawHud()
 
         # The following line copies our drawings from `canvas` to the visible window
-        self.window.blit(canvas, canvas.get_rect())
+        self.window.blit(self._canvas, self._canvas.get_rect())
         pygame.event.pump()
         pygame.display.update()
         return self
+
+    def _drawHud(self):
+        self.drawText(text=f"sensor {self._distance:.1f} m", location=(0, 0), color=(0, 255, 0))
+        return self
+
+    def drawText(self, text: str, location: Tuple[int, int], color: Tuple[int, int, int]):
+        """Draw a text"""
+        graphTxt = self._font.render(text, True, color)
+        textRect = graphTxt.get_rect()
+        textRect.topleft = location
+        self._canvas.blit(graphTxt, textRect)
+        return self
+
+    def _drawRobot(self):
+        """Draw the robot shape"""
+        trans = rotate(-math.radians(self._robot_dir)) @ translate(self._robot_pos) @ self._trans
+        shape = transform(trans, ROBOT_SHAPE)
+        pygame.draw.polygon(
+            surface=self._canvas,
+            color=ROBOT_COLOR,
+            points=shape)
+        return self
+
+    def _drawObstacle(self, location: ndarray, phantom=False):
+        """Draw the obstacle shape"""
+        trans = translate(location) @ self._trans
+        shape = transform(trans, OBSTACLE_SHAPE)
+        pygame.draw.polygon(
+            surface=self._canvas,
+            color=OBSTACLE_COLOR,
+            width=1 if phantom else 0,
+            points=shape)
+        return self
+
+    def _drawObstacles(self):
+        """Draw the obstacle maps"""
+        if self._obstaclesMap:
+            for i in range(0, self._obstaclesMap.num_obstacles()):
+                self._drawObstacle(location=self._obstaclesMap[i], phantom=True)
+        return self
+
+    def _drawSensor(self):
+        """Draw the sensor shape
+            
+        Arguments:
+        canvas -- the canvas
+        trans -- the transformation matrix
+        robot_location -- the robot location
+        robot._dir -- the robot direction (DEG)
+        sens_dir -- the sensor direction relative to robot (DEG)
+        """
+        angle = math.radians(90 - self._robot_dir - self._sensor_dir)
+        shape = SENSOR_SHAPE.copy()
+        shape = transform(rotate(angle) @ translate(self._robot_pos) @ self._trans,
+                shape)
+        pygame.draw.line(
+            surface=self._canvas,
+            color=SENSOR_COLOR,
+            start_pos=shape[0, :],
+            end_pos=shape[1, :]
+        )
+        if self._sensor_obstacle:
+            loc = self._sensor_obstacle
+            self._drawObstacle(np.array([loc.x, loc.y]))
+        return self
     
-def transMatrix():
+def _transMatrix():
     """Return the transform matrix to render the shapes"""
     #return scale((WINDOW_SIZE / WORLD_SIZE, WINDOW_SIZE / WORLD_SIZE)) @ rotate(-math.pi/2) @ translate((WINDOW_SIZE/2, WINDOW_SIZE/2))
     return scale((WINDOW_SIZE / WORLD_SIZE, -WINDOW_SIZE / WORLD_SIZE)) @ translate((WINDOW_SIZE/2, WINDOW_SIZE/2))
 
-def drawRobot(canvas:Surface, trans:ndarray, robot_location:ndarray, robot_dir:int):
-    """Draw the robot shape
-        
-    Arguments:
-    canvas -- the canvas
-    trans -- the transformation matrix
-    """
-    trans = rotate(-math.radians(robot_dir)) @ translate(robot_location) @ trans
-    shape = transform(trans, ROBOT_SHAPE)
-    pygame.draw.polygon(
-        canvas,
-        ROBOT_COLOR,
-        shape
-    )
-
-def drawSensor(canvas:Surface, trans:ndarray, robot_location:ndarray, robot_dir:int, sens_dir: int):
-    """Draw the sensor shape
-        
-    Arguments:
-    canvas -- the canvas
-    trans -- the transformation matrix
-    """
-    angle = math.radians(90 - robot_dir - sens_dir)
-    shape = SENSOR_SHAPE.copy()
-    shape = transform(rotate(angle) @ translate(robot_location) @ trans,
-            shape)
-    pygame.draw.line(
-        canvas,
-        SENSOR_COLOR,
-        shape[0, :],
-        shape[1, :]
-    )
 
 def scale(scale:Iterable[float]):
     return np.diag([scale[0], scale[1], 1])