+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue #10: Add save model
 - Issue #12: Simulated environment
+- Issue #15: Simulate obstacles
 
 ## [0.1.0] 2022-07-10